Split AppDelegate support code into focused files (#3104)
* Split AppDelegate support code into focused files

* Drain privileged installer output before waiting

* Clear pipe readers at EOF

* Avoid overlapping privileged output reads

* Deduplicate resolved Finder directories

* Harden shortcut routing edge cases

---------
