+++ Sources/App/ScreenIdentity.swift
@@ -1,0 +1,9 @@
+import AppKit
+
+extension NSScreen {
+    var cmuxDisplayID: UInt32? {
+        let key = NSDeviceDescriptionKey("NSScreenNumber")
+        guard let value = deviceDescription[key] as? NSNumber else { return nil }
+        return value.uint32Value
+    }
+}
