+++ cmuxTests/TerminalAndGhosttyTests.swift
@@ -1461,6 +1461,20 @@ final class GhosttyResponderResolutionTests: XCTestCase {
         override var acceptsFirstResponder: Bool { true }
     }
 
+    private final class DelegateTrackingTextView: NSTextView {
+        private(set) var delegateReadCount = 0
+
+        override var delegate: NSTextViewDelegate? {
+            get {
+                delegateReadCount += 1
+                return super.delegate
+            }
+            set {
+                super.delegate = newValue
+            }
+        }
+    }
+
     func testResolvesGhosttyViewFromDescendantResponder() {
         let ghosttyView = GhosttyNSView(frame: NSRect(x: 0, y: 0, width: 200, height: 120))
         let descendant = FocusProbeView(frame: NSRect(x: 0, y: 0, width: 40, height: 40))
@@ -1477,6 +1491,17 @@ final class GhosttyResponderResolutionTests: XCTestCase {
     func testReturnsNilForUnrelatedResponder() {
         let view = FocusProbeView(frame: NSRect(x: 0, y: 0, width: 40, height: 40))
         XCTAssertNil(cmuxOwningGhosttyView(for: view))
+    }
+
+    func testDoesNotReadTextViewDelegateForGhosttyResponderResolution() {
+        let textView = DelegateTrackingTextView(frame: NSRect(x: 0, y: 0, width: 40, height: 40))
+
+        XCTAssertNil(cmuxOwningGhosttyView(for: textView))
+        XCTAssertEqual(
+            textView.delegateReadCount,
+            0,
+            "Ghostty responder resolution must avoid NSTextView.delegate because AppKit exposes it as unsafe-unretained"
+        )
     }
 }
